+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Unidirectional Data-Binding of element `style` property
 - Support `a` as a formatting element under `p`
-- Todo MVC Sample
+- Todo MVC Sample: `require 'glimmer-dsl-web/samples/regular/todo_mvc'`
 
 ## 0.2.6